:art: update java sdk structure
