+++ javakafkademo/PUBLIC_SASL/pom.xml
@@ -30,8 +30,6 @@
     <build>
         <finalName>ckafka-operation</finalName>
         <plugins>
-
-
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
@@ -46,7 +44,6 @@
                     <compilerVersion>1.8</compilerVersion>
                 </configuration>
             </plugin>
-
         </plugins>
     </build>
 </project>
