+++ javakafkademo/PUBLIC_SASL/src/main/resources/client.truststore.jks
@@ -1,1 +1,1 @@
-## replace with real certificate .jks file
+## replace with real certificate .jks file## download [here](https://ckafka-public-certs-1255613487.cos.ap-guangzhou.myqcloud.com/ssl-certs/client.truststore.jks)
