+++ javakafkademo/VPC/pom.xml
@@ -15,7 +15,16 @@
             <artifactId>kafka-clients</artifactId>
             <version>0.10.2.2</version>
         </dependency>
-
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>1.7.5</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+            <version>1.6.4</version>
+        </dependency>
     </dependencies>
 
 
@@ -36,7 +45,6 @@
                     <compilerVersion>1.8</compilerVersion>
                 </configuration>
             </plugin>
-
         </plugins>
     </build>
 </project>
