+++ javakafkademo/VPC/src/main/java/ckafka/demo/CKafkaConfigurer.java
@@ -1,7 +1,5 @@
 package ckafka.demo;
 
-import ckafka.demo.defaultpoint.CKafkaProducerDemo;
-
 import java.util.Properties;
